bcore::impl_concat

Struct ConcatPostPrcess

pub struct ConcatPostPrcess {
    dataset: Vec<PostProcess>,
}

Fields
dataset: Vec<PostProcess>
Implementations

impl ConcatPostPrcess

pub fn new(folder: &[&str], root: Option<String>) -> Result<Self, ApiError>

pub fn get_time_end(&self) -> Result<Vec<f64>, String>

Retrieves the last time value from each dataset in the collection.

Returns

Result<Vec<f64>, String> - A vector of the last time values for each dataset or an error message if any dataset is empty.