bcore::impl_unique

Struct PostProcess

pub struct PostProcess {
    results: Results,
}

The PostProcess struct handles post-processing of simulation results.

It contains the path to the results folder, the root directory, and the processed results.

The struct uses the Results struct internally, which is expected to contain time and other data.

Fields
results: Results
Implementations

impl PostProcess

pub fn new(folder: &str, root: Option<String>) -> Result<Self, ApiError>

Creates a new instance of PostProcess.

Arguments

folder - The name of the folder containing the simulation results.

root - Optional root directory. Defaults to “./results/” if not provided.

Returns

Result<Self, String> - Returns the PostProcess instance or an error message if initialization fails.